First attempt to define repository for a single GDiffable value
This approach allows you to commit changes and rollback to any
previous state of the repo, but it is not possible to restore
it.

Possible issue:
Delta are one way and whether it is possible to mechanically invert
them is an open question.
Rollbacks are achieved applying all the patches from the empty
repository. If backwards deltas were also stored, it could be possible
to rollback "directly" from the current state.